requirements.txt
bcrypt==3.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
et-xmlfile==1.0.1
flash==1.0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gcloud==0.18.3
google-api-core==1.26.3
google-auth==1.30.0
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
loginform==1.2.0
lxml==4.6.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
numpy==1.20.2
oauth2client==4.1.3
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
protobuf==3.16.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.11
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==2.0.1
